Add Intellij helpers for spindle stubs and Lists.scala to Fsq.io. (#954)
* Add Intellij-configured Lists.scala.

If you use Lists.scala, you may have witnessed some issues during
type inference. This is due to a problem in upstream scala. For now,
we point Intellij at this specially configured version of Lists.scala
to provide type information in the place of the actual invoked code.

* Add copyright to Intellij xLists.scala

* Adjust header sizes

* Merge and remove the fingerprint from the output dir

Now this is just a copy from the results_dir, so it should be
legal to simply consume this from the artifact cache.

(sapling split of 91127e171fdc5599f9393038bbdecae5a127b278)