Add an extract arg to RemoteSource and uncouple from Rpm system. (#1376)
* stash extract arg for remote source

* Add an extract arg to RemoteSource and uncouple from Rpm system.

This allows any target to depend on a versioned file that is
hosted in a configured site.

This allows for extracting the download and then adds to the
product graph as a resource. This would be useful for
bootstrapping files to the classpath, although the upcoming
use case is actually something we only want to be local, so still
figuring that part out.

The remote_files tasks and subsystems should be completely broken
out of the Rpm module, will do that as a followup. This also lacks
any testing, because it predated the test infra - I hope to add
a testproject and beginning of a test suite when I write the followup.
I know that there is some error handling that is needed to catch
incongruitites and failures in the extract pipeline.

* Update some comments

* fix vt API bug

* review feedback

* add extract as a payload field

(sapling split of 9a55072fb6a5be8faa8f997472a77d5096a62411)