Remove clean from pants task setups. (#1277)
* Remove clean from pants task setups.

Tasks that might leave partial results in a workdir were historically
left to the task to clean. But that became bugged when the workdirs
were moved to a symlink, since the clean algorithm didn't have symlink
awareness.

The consequence of this was that calling clean=True would silently overwrite
the symlink with a real dir and break the cache symantics.

For resolves, the fresh workspaces was still needed. I changed upstram
Pants to clean workdirs by default, so any failure now cleans.

https://github.com/pantsbuild/pants/issues/4051
https://github.com/pantsbuild/pants/issues/4137

* Fix linter and missed variables in the refactor

(sapling split of c8a374715e5d7eb31eac507c70a3fe9db476af01)