Add upkeep scripts to fetch and package jpostal/libpostal blobs. (#1500)
* Add upkeep scripts to fetch and package jpostal/libpostal blobs.

I understand that this library might become updated soon, and since
this requires setting up build pipelines in so many domains, I went the
extra mile and wrote up some packaging scripts.

To update either library, you can update the version in the env.sh file
and run `./upkeep run package-jpostal.sh`.

That will fetch the source code, compile libpostal and the shared object
files, and create a tarball that holds them. The tarballs are output to
dist with a specific namespace, they need to be uploaded to bodega for the
org to consume. The namespacing is significant, upload directions are in the
script.

A followup script will add one more upkeep script that consumes the packaged
libs, which enables interacting with jpostal from the repl and unit tests.

The last thing needed is to rebuild the RPMs and upload them to the repo, which
will enable production hadoop machines to do the same. See the readme in the
matching src/redhat folders.

* Add forked libpostal to get release tag.

Guard added to unfork next update.

This was also checked against devbox environment, so it should be okay
for both OSX and linux.

* Adjust further for linux.

Add the datadir extracted target as well. This is confirmed
working for devboxes as well.

Remaining work:
* Upload the blobs to bodega
* test the centos6b blobs on centos7 and see if they work
* adjust data dir and publish final rpms
* upload rpms to bodega repo
* test full install on 4sq-test fleet
* add RPMs to hadoop fleet
* figure out linux crash when unpacking remote_sources
* figure out how to add libpostal_data target as repl dependency
* add library.path to repl jvm invocation
* add upkeep task to pull down blobs per-OS.

That looks like a lot (and it is) but all the big steps have finally
been figured out, written, and validated.

(sapling split of 14e00226c53f957aeaedf33356b9100243bee464)