Bootstrap jpostal blobs through upkeep (#1527)
* Add upkeep scripts to fetch and package jpostal/libpostal blobs.

I understand that this library might become updated soon, and since
this requires setting up build pipelines in so many domains, I went the
extra mile and wrote up some packaging scripts.

To update either library, you can update the version in the env.sh file
and run `./upkeep run package-jpostal.sh`.

That will fetch the source code, compile libpostal and the shared object
files, and create a tarball that holds them. The tarballs are output to
dist with a specific namespace, they need to be uploaded to bodega for the
org to consume. The namespacing is significant, upload directions are in the
script.

A followup script will add one more upkeep script that consumes the packaged
libs, which enables interacting with jpostal from the repl and unit tests.

The last thing needed is to rebuild the RPMs and upload them to the repo, which
will enable production hadoop machines to do the same. See the readme in the
matching src/redhat folders.

* Add jpostal_blobs to upkeep for libhadoop

Easiest to colocate, imo.
This also preps the RPMS

* update fs-bootstrap path

* Properly unzip libpostal

* Add upkeep scripts to fetch and package jpostal/libpostal blobs.

I understand that this library might become updated soon, and since
this requires setting up build pipelines in so many domains, I went the
extra mile and wrote up some packaging scripts.

To update either library, you can update the version in the env.sh file
and run `./upkeep run package-jpostal.sh`.

That will fetch the source code, compile libpostal and the shared object
files, and create a tarball that holds them. The tarballs are output to
dist with a specific namespace, they need to be uploaded to bodega for the
org to consume. The namespacing is significant, upload directions are in the
script.

A followup script will add one more upkeep script that consumes the packaged
libs, which enables interacting with jpostal from the repl and unit tests.

The last thing needed is to rebuild the RPMs and upload them to the repo, which
will enable production hadoop machines to do the same. See the readme in the
matching src/redhat folders.

* Add jpostal_blobs to upkeep for libhadoop

Easiest to colocate, imo.
This also preps the RPMS

* update fs-bootstrap path

* Properly unzip libpostal

* move library.path for repl to libhadoop

* Force libhadoop

* fix typo

(sapling split of 3916a6873dadd42941fad79ed61cf50dc6f3be10)