Unify the libpostal version to not use the upstream `v` nomenclature. (#1536)
GitHub does not consider it part of the the version, so neither should we.

This also moves the version in the BUILD file to consume the same environmental
variables as the rest of the scripts. That leaves the version defined in
a single place, the env files

There should probabl be a Pants helper for this at some point, but
this is fine for now.

(sapling split of 68e4cc42fe0883bf323ce9aca4f99f5e72bc5933)
(sapling split of d4c9c3f86276dd5b591b2251175a19e879b213c8)