Fix '../current/<task>: No such file or directory' in upkeep fails. (#1733)
This happens in the cases where someone has a fresh checkout or if
they have totally deleted the dependencies/ folder.

It is actually not a failure here, but it does obscure the
actual issue by closely resembling a error trace.

The cause of this issue was already known and the fix was
applied. The problem was that the implementation had a
very very very very subtle bug.
(sapling split of 3805ba6a0847d0af4151df4a80297f2d753cf089)