ivy/hadoop: fix classpath conflict that busted up Application Master UI (#1948)
The web UI exposed by the Hadoop Application Master for a job uses Jetty as the
servlet container and Jersey as the web framework. As part of its use of Jersey,
the Hadoop AM code uses the javax.ws.rs-api jar to process XML. The Jersey jars
on the CDH 5.4 classpath requires a certain version of this jar.

Unfortunately, the CDH 5.4 classpath has multiple versions of Jersey and multiple
versions of this XML jar. And the ordering of the classpath is non-deterministic.

Solution: Unexclude this jar to reintroduce it to our Hadoop classpath, and
because we put our jars first on the classpath, that version will take precedence
on the classpath.

Did I say I *really* *love* the classpath?
(sapling split of ba9c0261e37e9fe366f64f8ce801d7f098e971a6)