Remove TestSerial annotation and remove junit-runner dep. (#2540)
Pants bootstraps its own junit-runner - the only reason we had
this on the classpath was so that we could use the Annotation.

But it turns out that the annotation is ignored at the method level.
This has the practical outcome of accidentally disabling Unit tests.

Changing

@test
def testMeOften()

to

@testserial
def testMeOften()

is a noop from the runner, but also has the side-effect that regular
junit no longer considers this a test case. So forcing serial was
actually just forcing a skip.

To avoid regressing here, we are removing the runner as a valid 3rdparty
dependency - the only supported way of forcing serial junit is to add
the 'concurrency' field to the BUILD file.
(sapling split of b09268cd2f5893c249ce2dea1ddfa4e207a07372)