Remove rmtree for invalid target roots. (#2838)
Not 100% sure that this is the fix here. We added an 'extracted' field
that makes this hard to reason about.

The proper answer here is to go forward with the target.id or bite the
bullet and allow it to consume from .pants.bootstrap.

The issue here is that several of these cached files are used
at runtime internally - so we consume from a symlink in order
to not thrash the compile crash. But I think this could use a
rework - I added it under extreme time crunch and could
reexamine the logic.
(sapling split of 43bba08fed1ac992065476eb2399bdddc62a87d9)