Quiet some upkeep chatter (#2972)
Upkeep started reporting failures 2x, since in almost (all?)
cases, the task itself exited uncleanly with a message. That
exit code is actually what triggers the trap.

It might be a better API to have the task push a message and
leave the exit code to the executor. But in abscense of an appetite
for that work, I am just removing this one for now.

The second piece is a guard around the kestrel download,
specifically squashing the progress bar on Jenkins runs.
Jenkins streams to console, so progress bars can end up
totally dominating the CI output.
(sapling split of a2fae6bef49474c360308c8510d9a02de055ea94)