Move all spindle.codegen and supporting shaded code under a monolithic root. (#3610)
The impetus for this change was the upcoming patch that converts the repo from our custom PomPublish back to the upstream Pants publish.jar.

In the old world, we would have had to configure a library that help just the shaded versions
of the code. This just makes the boundaries of the shaded code clearer, and makes for a more
intuitive publish story,

Ideally, no outside targets would ever be allowed to reach into the codegen library for that
shaded code. We have some migration tests that we should maybe consider disabling as
the migration effort has been all but abandoned FWICT.

This removes the build files from the submodules in an attempt to isolate the shaded code.
This makes the targets less modular than they were previously. But outside of commits
meant to improve the build process, I only see 6 true spindle.codegen changes in the last year.
So the cost is minimal on that front.

This also adjusts the runtime deps and only insists on things not brought along by the codegen binary. I believe this to be correct but would be willing to second guess that choice.
(sapling split of 0682971a0ce4ef0ea1de985efbba7669ecd7f693)