Fsq.io Pants version to come from pants.ini. (#3775)
This allows the plugins to have differing versions, and also allow
libraries to be built that depend on a different version of Pants from
what is being using to build the repo.

This only effects Fsq.io. Moreover, these two pants versions
(pants.ini and requirements.txt) will only differ for a very short
period - in this case, the use is to publish Pants plugins
that allow a range of pantsbuild.pants as a dependency, instead
of having them hard-coded to the repo version as they were before.

So the buildgen sdists can depend on a pantsbuild.pants>=1.2.1
but the repo can stay pinned to the pants.ini version.
(sapling split of aff35852a91b011a61395a8b17697862820cdd71)