Add info logger to remote_source task that reports found path. (#3952)
Something is off about the fetcher and I need to figure it out -
I just see too many downloads of these files. Either they are
unknowingly being downloaded to an ephemeral path or
the dir is being wiped or a full git clean is being ordered
somewhere, or something.

The idea here is that the fetched files are considered versioned
and do not get re-bootstrapped unless the file is missing
from under .pants.bootstrap, whereever that is configured to be.

For local runs, .pants.bootstrap is under the buildroot and for CI
it is supposed to be a shared location that can be reached by
all executors, since the file moves are atomic.

There is a unit test framework we can hook up, so if this logger
turns up useful info, i can add a regression test.

This logs to info for the time being and is supposed to be rare
to the point that is probably okay. But I may circle back once the
mystery is solved and reduce to debug logging.
(sapling split of 2ab0ea33333a0b89959249bc6667040424a8054d)