Make reading from cache the default ini setting and set per-goal. (#3950)
Do not sweat the correctness of this patch too much - as soon as
this upgrade bakes a day or two, I intend to turn cache read/writing
on by default and delete almost all of the cache entries in all .ini files.

The opt-out of caching is a remnant of earlier Pants days, when
tasks had to manage their own reading and writing and workspace
management. Caching is enabled in tasks by default - tasks that are
not intended to write to the cache now declare that as task
properties and do not require management by the consuming repo.

I always stumble over tasks I thought were writing to the cache
and were not, etc.

I will motivate that choice in the upcoming patch. In this commit,
there are a few things of potential note, but it is mostly
intended to consolidate config in advance of that.

Interesting things in this set:
  - cache _reads_ are on by default. This means that anything that
    can access a remote cache considers those artifacts useful.
    Since we configure each task to write to the cache individually,
    I think we can consider that assumption to be active in the current
    world and that we would want to know where it was untrue.
  - compile.zinc had local cache writes disabled. It was passively
    reading from the workdir but not the cache. This is possibly
    to conserve space on the local disk.
    - I have enabled the local writes here but limited to two artifacts
      per target.

As said above, I plan to delete most of these options sometime this
week, this is just a staged rollout that hopes to suss out
bad citizens in-advance of that move, and the test caching change.

I also am strongly considering a shared local cache, for local
and CI invocations. The current system is only really useful to
protect against poisoned cache artifacts, but we no longer
believe those are live, and would prefer to enable/find/fix instead
of being overly defensive.
(sapling split of b12eb9eecb93800e7ad887f2f69820dadfc85118)