Pull out a base class for target bag injection. (#4021)
I have several implementations of this scattered among
our Pants backends, but this is the canonical one.

This change pulls out the jvm logic from the target bag
and tightens the API. It also includes a really nice perf
bump since it no longer instantiates any dependent
targets but instead traffics purely in unhydrated Addressables.

The motivation for this change is the upcoming move to the
upstream Python backend, where I am considered using a
"global classpath" for pex resolution as well. Not sure
how that will work exactly but breaking out the base class
was the prerequisite.

Buildgen and buildgen target bags should probably be migrated
to implement this where appropriate, but that is like a
fixit-level priority.
(sapling split of bdaf93cf0b3ac5e71d0587b9c316ce1a3e5a6dc3)