Port Quiver to the monorepo (#4398)
This has been tested to work but is not yet deployed. Some important notes:

- Go at build time is bumped to 1.9.3
- Go-thrift is bumped to .93
- The version is not properly burned into quiver

The version bumps were a requirement of regenerating the thrift. There
is no version of thrift and Go that worked with the quiver codebase outside of the
above. The github version of quiver was actually building with checked in
generated Go, which is how that slipped by.

The last item needs to be solved before we can deploy the new loko .
As configured, the version is set with LDFLAGS and imported into
main.go, which has proved troublesome. There is a build-flagss 
Pants flag but I have had limited success. 

Pushing now to avoid merge conflicts and will have to update and deploy
this loko package.

(sapling split of d47d208af8ba9df002135269ee910994126551f5)