Import country_revgeo shapefiles jars into Fsq.io (#5296)
* Fix the remote_sources task runtime_classpath.

This is almost certainly broken for our internal usecases as well.
The Pants 1.3.0 upgrade likely is the culprit, which introduced some
new classpath products mechanics.

This also simplifies the remote_source task a bit. It used to set a
stable_symlink with a path that wouldn't invalidate compiles. But
for Pants 1.4.0 we will get stable task results_dir paths by
default, so the complexity isn't worth inlining in this task.

* Move cc-shapefiles quadtree library into Fsq.io

* Finish moving into fsq.io

* fixlint

* buildgen

(sapling split of 77de0d8f939f8d6f294801985dfc2f58996b33cd)