Consolidate all jvm_options into pants.ini and remove from internal. (#6176)
This was one of the most confusing things about Fsq.io, finding out
what setting was actually the one you wanted in prod. This moves
all the jvm config into pants.ini where it can be safely used internally
or elsewhere.

We have some file paths that we want available to our jvm - native
extensions and so on, and those are now being set in the env.sh file.
This is a much better fit, since those are the scripts that are populating
those paths.

This has the added effect of being easier to reason about from the CLI.
The way I tested this was to run `repl`, `test`, `compile` and `run` and examine
the invocation. But you can see any optionm that is actively set by running
`./pants options --name=jvm_options` or `./pants options --scope=jvm`

This patch was needed for some native lib integration that we want internally
and also to fix Fsqio jvm invocations, which we only being set for a subset
of JVM calls, which resulted in heap starvation at test time and other pathologies.
(sapling split of 2fa1835d29d9871b64c1b93719fff93a479591cb)