Remove the Yourkit libs from the Pants debug args. (#6528)
The Pants code (and the constructed java cmd) looks like this
should work - but I think that it is being interpreted as an arg
or something. Something about this makes the java runtime think that
we are passing it as a class.

I opened a ticket to return to this and productionize Yourkit
once and for all. I would really like to make perf and debugging
simpler so that we lose less time to people looping for
printlns.

This also moves the default debug port option to the internal
env.sh - why force this on our friendly OSS users - it is
a Foursquare quirk.
(sapling split of ca00196d9514ab9b80171c9e45d152760cb7f53d)