Update import sorter to be scalafmt-compatible (one per line) (#7127)
Turns out that scalafmt only sorts groups of imports within the curly braces, and we also lost import de-duping. This reintroduces the classic ScalaImportSorter, but wired through a one_per_line argument that makes it format the output the same way as scalafmt. So ScalaImportSorter runs and does all the sorting, and outputs the imports in the format that scalafmt expects, which then is presumably a no-op on the imports.

* Update import sorter to be scalafmt-compatible (one per line)
* Add extra arg to ScalaImportSorter in fs bag
* Add commandline options to main classes that use ScalaImportSorter
* Try to clean up the import formatting code
* Move ScalaImportRewriteRule to scala_imports to avoid cycle dep
* MORE TYPE HINTS
* Fix all the mypy warnings in source_code_analysis
(sapling split of 580db959d16572462494ceddc40be6c2b2963249)