have rogue catch MongoInterruptExceptions and add details (#9178)
This adds a more detailed error message with some additional context
and the current thread name to help with debuggging, so hopefully the
next person to hit this won't have to spend two months tracking down
the root cause.

As mentioned in the comment, there are some shenanigans happening
here with string manipulations to generate this error message. This
was mostly motivated by me not wanting to stick a 200 character line
in this file, but I claim it's also slightly more efficient than any
of the less custom ways this could have been written. And at least
it's quarantined to its own helper function.

This commit also adds some basic tests for the twitter Future
callbacks, and the new exception encoding.
(sapling split of c402958f499dfc1fd025911cc22c51e8a0003e36)