Don't hard fail for newer OS (#9646)
I got an OSX update notification, so odds are good someone
is gonna be in this spot soon.

The proper upgrade path used to be making sure the binaries are all copied
to the proper 10.x namespace. This is two steps, and probably should be 3:
* Upstream Pants needs to rebuild 10.14 binaries and host them (happening soon enough)
* Copy 10.13 binaries to 10.14 namespace on bodega
* Symlink our 10.13 binaries to 10.14 namespace (artifactory doesn't support symlinks)

But with the Pants 1.7 upgrade, we started
supporting the upstream "scripts" paradigm, which is essentially the same
as our internal 'remote-sources' concept but supports platform independent
blobs, so instead of

bin/node/mac/10.13/v6.90/node.tar.gz
  it becomes
bin/node/v6.90/tar.gz

That is a hell of a lot easier and saves us a ton in bandwidth and storage costs,
especially on our hosted artifactory instance.

I was able to convert the remote_sources backend to support this new scripts setup,
currently only being used by cc-shapefiles source blobs.

TLDR: We should move the internal remote-sources blobs to the scripts namespace
instead of 10.14 but this hack buys us time to do that.

(sapling split of edcf1e54aee8abb16045e62097ba7e77a4f45b5f)