remove maxScan from rogue queries (#11738)
This is a deprecated option in mongo itself, and we haven't been setting it
on queries since the java driver upgrade (which is what broke the tracing
services where maxScan was being used as a query limit). Fortunately it
wasn't in widespread use and easily removed -- there are a few places where
I replaced it (the aforementioned tracing query) but otherwise I've just
dropped it as unnecessary in most places.
(sapling split of 836ce066cc2474b42c76e69da66c8604c80e53d3)