exceptionator: fix running with custom mongo server location (#11870)
We lost exceptionator's mongo boot code in the thrift switchover,
which I missed in the review (it was a large change!) but discovered
when trying to deploy a bug fix last week. This restores it and
updates it for the new world, including adding back some indexes that
were missing on the thrift models.
(sapling split of 4a42fbb5e39c33c5c3e5967ec04e9785660cc122)