Fix Pants Publish through Artifactory. (#13596)
This is not working yet. When publishing a dummy jar library
to Artifactory:
  - using basic auth and maybe/not using the netrc
      - disabling the `auth` in the repos in ini is way better for debugging
  - Does not include the PGP signing for jars
  - 405s

But all the pieces are now back in place from the Pants side.

The remaining work is for an Artifactory admin to figure out how
   - The correct Artifactory repository to publish to
   - How to auth with netrc or without
   - Determine how to add jar signing config
       - I included commented out example code that crashed during the
         signing attempt - it may just need adjustment
   - Artifactory permissions for privileged publishing users

I think that we are basically almost there and someone closer
to Artifactory fluency could prove it out in less than a day.

We are preparing to publish Scala 2.12 jars and there are other
use cases for publish in the air, so this should be picked up and
completed if we can find a willing party.

(sapling split of 71e30b0838b63069de14e02a9b2285f0b7e4d320)