Adjust the cloudera jar repo used by Fsq.io (#14378)
Even though we pin to pure mortbay 6.25, during ivy resolution
ivy still checks the transitive dependencies (and therefore  poms)
for versions outside that pin discovered in the transitive
dependency graph.

Normally this is invisible to us, but we are seeing Ivy crash when
resolving a cloudera-classifier jar of mortbay.jetty

```
                     	  -- artifact org.mortbay.jetty#jetty;6.1.26.cloudera.4!jetty.jar:

                     	  /Users/mateo/.m2/repository/org/mortbay/jetty/jetty/6.1.26.cloudera.4/jetty-6.1.26.cloudera.4.jar

                     		::::::::::::::::::::::::::::::::::::::::::::::

                     		::          UNRESOLVED DEPENDENCIES         ::

                     		::::::::::::::::::::::::::::::::::::::::::::::

                     		:: org.mortbay.jetty#jetty-util;6.1.26.cloudera.4: not found

                     		:: org.mortbay.jetty#jetty;6.1.26.cloudera.4: not found

                     		::::::::::::::::::::::::::::::::::::::::::::::

```

I guess the either removed that releases repo or it never existed and the caching fooled
us all. Wouldn't be the first time.
(sapling split of 73c6d1c918872f33de624ac8e7387c92333256c5)