Fix many mypy type errors. (#19499)
* Fix many mypy type errors.

At the start of development of this patch, there were two mypy errors that
caused mypy to terminate early. Fixing those, we find that mypy reports
220 errors in our codebase.

```./pants mypy src/python:: | grep "error: " | wc -l
220
```

Most of those errors fall into two categories that I am not addressing in
this patch. The first is `Definition of "some_method" in base class "Class1"
is incompatible with definition in base class "Class2"`. In our codebase,
this usually means that some `Class3` (note that the error message does not
name `Class3`) inherits from both `Class1` and `Class2`, and that one of
these classes has an `@abstractproperty` and the other class implements that
property as an instance variable. (See:
https://foursquare.slack.com/archives/C2K9V6NLE/p1576190434023700). I don't
think this should be an error, and Mathieu thinks it might be fixed in newer
versions of mypy, so I am not handling it now. There are 134 instances of
this error at present.

The second type of error is `Read-only property cannot override read-write
property`. This error seems to happen when a property is read / write only
so that it can be set shortly after object instantiation (for example, a
luigi parameter), and then a subclass wants to override it with a read-only
implementation. I think the only solution in this case is to implement a
setter method in the sublass that just passes. But I have not done so at this
time because the solution is ugly and offends my aesthetic sensibilities.
There are 9 instances of this error at present.

That leaves, before this patch, 77 mypy errors.
```./pants mypy src/python:: |
  grep "error: " |
  grep -v "is incompatible with definition in base" |
  grep -v "Read-only property cannot override read-write property" |
  wc -l

77
```

And after this patch:
```./pants mypy src/python:: |
  grep "error: " |
  grep -v "is incompatible with definition in base" |
  grep -v "Read-only property cannot override read-write property" |
  wc -l

0
```

I have tried to avoid making an logical changes to the code itself, or
really any changes at all to the code itself, but this has not been
possible in every instances.

(sapling split of 715068d22275d422eb04113e7b9aa805c27e8670)