pull finagle-kestrel into foursquare.web (#20844)
Finagle removed its kestrel queue client in version 6.45.0. In order to upgrade,
we were forced to either consume a shaded version of its legacy jar or pull it
into our codebase and maintain it ourselves going forward. We did not find
success on the shading route (though certainly not for lack of trying), but
fortunately were able to upgrade the kestrel client for a modern finagle version
without too much trouble (to be shipped with the finagle 20.1.0 upgrade).

This only pulls in the pieces of finagel-kestrel we use, and their associated
tests. All of this code is pulled in more or less as is, modulo some package
structure and linting changes.

(sapling split of e8fbb3ebfb71251e0591b2f94d3bb47188006bb7)