upgrade the ammonite scala repl to 1.6.7 (#23140)
This is the last version supported on scala 2.11. For the most part this was a
drop in replacement, save for the fact ammonite no longer accepts multiple
instances of the predef script arg. This change works around that by passing
the default script as code (it currently just sets the repl prompt to the run
mode) but we perhaps want to figure out a better solution moving forward.
(sapling split of 89ad780a1b1476c3b51359ff3d79c1cb2f1e0df0)